Find sensitive data via regex rules
The script makes the requests to the domains and subdomains in the list with the maximum depth to 5 recursively. All the domains, subdomains are uniquely added to the list called "visited". Afterwards, when one of the URL in the list called "visited" is visited and is matched if one of the regex patterns in the list called "patterns", the domain or subdomain are added to the list called the_final_urls, skips if the domain or subdomain that have already been visited.
Simply, it finds sensitive data via regex rules in the URL adressess which provided by you.